Implementa lógica de likes e comments do postsId
