+++ src/app/components/post_list/styles.module.scss
@@ -3,6 +3,7 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  padding-bottom: 50px;
 }
 .titleContainer {
   font-size: 15px;
