+++ src/app/posts/[postId]/styles.module.scss
@@ -3,7 +3,7 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  padding-bottom: 20px;
+  padding-bottom: 50px;
 }
 
 .callAction {
