--- src/hooks/useDeviceDetection.ts
@@ -1,22 +1,0 @@
-// hooks/useDeviceDetection.ts
-'use client';
-
-import { useState, useEffect } from 'react';
-
-export function useDeviceDetection() {
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const checkMobile = () => {
-      const mobileRegex = /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i;
-      setIsMobile(mobileRegex.test(navigator.userAgent));
-    };
-
-    checkMobile();
-    window.addEventListener('resize', checkMobile);
-
-    return () => window.removeEventListener('resize', checkMobile);
-  }, []);
-
-  return { isMobile };
-}
